# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "タスみる API",
    "description": "仕様書からタスクを自動生成し、担当候補を提案するAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Status",
        "operationId": "get_job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/error": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Error",
        "operationId": "get_job_error_api_jobs__job_id__error_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/result": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Result",
        "operationId": "get_job_result_api_jobs__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalProjectOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/requirements": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Requirements",
        "operationId": "get_job_requirements_api_jobs__job_id__requirements_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/tasks": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Tasks",
        "operationId": "get_job_tasks_api_jobs__job_id__tasks_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/dependencies": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Dependencies",
        "operationId": "get_job_dependencies_api_jobs__job_id__dependencies_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/validation": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Validation",
        "operationId": "get_job_validation_api_jobs__job_id__validation_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/assignments": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Assignments",
        "operationId": "get_job_assignments_api_jobs__job_id__assignments_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FinalAssignment"
                  },
                  "title": "Response Get Job Assignments Api Jobs  Job Id  Assignments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DependencyDocument": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/Dependency"
            },
            "type": "array",
            "title": "Dependencies"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/backend__pipeline__dependencies__schema__ValidationIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "graph": {
            "additionalProperties": true,
            "type": "object",
            "title": "Graph",
            "description": "DependencyGraph.to_dict()の結果"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At",
            "description": "ISO8601形式の生成時刻(UTC)"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "dependencies"
        ],
        "title": "DependencyDocument",
        "description": "dependencies.json の全体"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "FinalAssignment": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "assigned_member_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Member Id"
          },
          "decided_by": {
            "type": "string",
            "enum": [
              "ai",
              "human"
            ],
            "title": "Decided By"
          },
          "overridden": {
            "type": "boolean",
            "title": "Overridden",
            "default": false
          },
          "override_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Reason"
          },
          "ai_recommendation": {
            "$ref": "#/components/schemas/AssignmentResult"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "decided_by",
          "ai_recommendation"
        ],
        "title": "FinalAssignment",
        "description": "人間による最終的な割り当て決定。\n\n`ai_recommendation`にAIの推薦をそのまま保持することで、\n「AIの推薦」と「最終的な人間の割り当て」を常に分離して追跡できる。"
      },
      "FinalProjectOutput": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectInfo"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "type": "array",
            "title": "Requirements"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array",
            "title": "Tasks"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/Dependency"
            },
            "type": "array",
            "title": "Dependencies"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/Member"
            },
            "type": "array",
            "title": "Members"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/FinalAssignment"
            },
            "type": "array",
            "title": "Assignments"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationSummary"
          },
          "metadata": {
            "$ref": "#/components/schemas/ProjectMetadata"
          }
        },
        "type": "object",
        "required": [
          "project",
          "requirements",
          "tasks",
          "dependencies",
          "members",
          "assignments",
          "validation",
          "metadata"
        ],
        "title": "FinalProjectOutput",
        "description": "依頼のFINAL JSON節に対応する、最終的な構造化プロジェクト出力"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "team_id": {
            "type": "string",
            "title": "Team Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "project_id",
          "team_id",
          "status",
          "progress",
          "created_at"
        ],
        "title": "JobStatusResponse"
      },
...
      "ProjectInfo": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "exported_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exported At"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "ProjectInfo",
        "description": "`project`セクション。仕様書自体の内容は含まない（Requirement/Task等に\n既にすべて含まれている）。"
      },
...
      "Requirement": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^REQ-\\d{3,}$",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "system_purpose",
              "target_user",
              "functional",
              "non_functional",
              "constraint",
              "assumption",
              "deliverable",
              "technical",
              "business_rule"
            ],
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low",
              "unknown"
            ],
            "title": "Priority",
            "default": "unknown"
          },
          "origin": {
            "type": "string",
            "enum": [
              "explicit",
              "inferred"
            ],
            "title": "Origin",
            "default": "explicit"
          },
          "source_reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "description",
          "confidence"
        ],
        "title": "Requirement",
        "description": "最終的な構造化要求（requirements.json の1要素）"
      },
      "RequirementDocument": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "type": "array",
            "title": "Requirements"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/backend__pipeline__requirements__schema__ValidationIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At",
            "description": "ISO8601形式の生成時刻(UTC)"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "requirements"
        ],
        "title": "RequirementDocument",
        "description": "requirements.json の全体（1つの仕様書に対する抽出結果）"
      },
...
      "Task": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^TASK-\\d{3,}$",
            "title": "Id"
          },
          "requirement_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirement Ids"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low",
              "unknown"
            ],
            "title": "Priority",
            "default": "unknown"
          },
          "estimated_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Hours"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "acceptance_criteria": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Acceptance Criteria"
          },
          "source_reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "related_sources": {
            "items": {
              "$ref": "#/components/schemas/RagSource"
            },
            "type": "array",
            "title": "Related Sources"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review",
            "default": false
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "confidence"
        ],
        "title": "Task",
        "description": "最終的な構造化タスク（tasks.json の1要素）"
      },
      "TaskDocument": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array",
            "title": "Tasks"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/backend__pipeline__tasks__schema__ValidationIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At",
            "description": "ISO8601形式の生成時刻(UTC)"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "tasks"
        ],
        "title": "TaskDocument",
        "description": "tasks.json の全体（1つの要求ドキュメントから分解されたタスク一覧）"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationReport": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "missing_requirements": {
            "items": {
              "$ref": "#/components/schemas/MissingRequirement"
            },
            "type": "array",
            "title": "Missing Requirements"
          },
          "duplicate_tasks": {
            "items": {
              "$ref": "#/components/schemas/DuplicateTaskGroup"
            },
            "type": "array",
            "title": "Duplicate Tasks"
          },
          "dependency_errors": {
            "items": {
              "$ref": "#/components/schemas/DependencyError"
            },
            "type": "array",
            "title": "Dependency Errors"
          },
          "workload_warnings": {
            "items": {
              "$ref": "#/components/schemas/WorkloadWarning"
            },
            "type": "array",
            "title": "Workload Warnings"
          },
          "skill_mismatches": {
            "items": {
              "$ref": "#/components/schemas/SkillMismatch"
            },
            "type": "array",
            "title": "Skill Mismatches"
          },
          "constraint_violations": {
            "items": {
              "$ref": "#/components/schemas/ConstraintViolation"
            },
            "type": "array",
            "title": "Constraint Violations"
          },
          "task_quality_issues": {
            "items": {
              "$ref": "#/components/schemas/TaskQualityIssue"
            },
            "type": "array",
            "title": "Task Quality Issues"
          },
          "assignment_score_anomalies": {
            "items": {
              "$ref": "#/components/schemas/AssignmentScoreAnomaly"
            },
            "type": "array",
            "title": "Assignment Score Anomalies"
          },
          "workload_summaries": {
            "items": {
              "$ref": "#/components/schemas/WorkloadSummary"
            },
            "type": "array",
            "title": "Workload Summaries"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At",
            "description": "ISO8601形式の生成時刻(UTC)"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "ValidationReport",
        "description": "依頼のRESULT節に対応する検証結果全体"
      },
...
      "backend__pipeline__requirements__schema__ValidationIssue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "requirement_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirement Ids"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ValidationIssue",
        "description": "Requirement Validationステージが検出した問題点"
      },
...
    }
  }
}